automatically detect GoPro timelapse videos during processing
GoPro timelapse videos are taken at an interval of one picture every
few seconds. This information is not properly encoded in the video
stream which is fixed to 25 FPS. Luckily the actual recording duration
is available in a metadata stream the camera attaches.

The timing is not quite accurate: It appears the last picture taken
always has a "view time" of 1s, regardless of the interval. For some
test videos I have the following numbers:
interval 2s, 58 stills, expected: 116s, reported: 113s
interval 5s, 56 stills, expected: 280s, reported: 280s
Since we don't know the interval we can't properly account for this,
so we just round up to the closest sensible interval.